[#112]: Move common functions to utils file
